requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.24
psycopg[binary]==3.2.13
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.6.0
